Generating core module from DSL (Closes: #7).
A really ugly implementation of the core generator. I didn't have the
time to experiment with pymeta inheritance (which I don't even know if
it is implemented), so I copied the base memetalk grammars to create the
core.h and core_tr.g, which adds a syntax for "object" declarations,
necessary to express Behavior and Object.

-The core module written in memetalk is in src/core.md.

-The core module generator is src/gen_core.py. It creates the source
 file core_module.py, which is loaded from i.py.

-Some prims where removed since they could be expressed in memetalk.

-Also, fixed a bug when parsing methods with alpha names (like <=)

All tests (except eval1, which seems to be broken for reasons unrelated
to this commit) passed.