Continuing execution when closing debugger (Closes #50)
To be able to catch `QWidget`'s close event, we need to have that python
object with the method `closeEvent`. The way I did this was to create an
intermediate `_QMainWindow` class with `closeEvent()` inheriting from
`QMainWindow`. Whenever `closeEvent` is triggered, that object checks if
the corresponding memetalk object has the method `closeEvent` calling it
if affirmative.

During this development, I noticed `super_ctor_send()` was running the
super method with delegate as the receiver -- and the approach above
need a hold on the outtermost (rp) instance. So, insteand of `receiver`,
we are passing `instance` to `setup_and_run_fun`.

Finally, I refactored `prim_basic_new()` (which is not really a
primitive) making it a `interpreter.create_instance()` call.