Stepping execution on var access.
Debugger is already able to change variable values in context. I was
mislead into thinking it didn't because, without pausing there, as soon
as the debugger paused on an expression involved the changed variable
(which already used `eval_access_var` to fetch its value), the old value
is the one about to be used on the paused expression.

Closes #55.