Working around some closure bug.
calls to enablePluginsWith closure seems to be resetting "variables" for
all closures inside that are passed to ExplorerEditor ctor.

To reproduce, declare a variable inside a ModuleExplorer editor (which
should now be visible in all the editors there), and add a new function
or method. Now the var does not exist anymore.