Debugger can be invoked more than once (Closes #72).
when "continue"ing, instead of the coroutine back and forth, we can
simply destroy the window, exit the loop and return the command
"continue" from the main dbg function, which will be received in the
other end.

In that way, the debugger coroutine is properly finished and can be
started again, where previously it's coroutine would be stuck in
"prim_vmprocess_continue" waiting for the end of a possible qt slot to
be switched back and cleaned up, which may never happen.

Oh yeah, and now I can debug the debugger!