improving eval_do_send_or_call.
Because "throw_with_value" may not throw an exception (and instead,
start the debugger) the rsulting 'fn' may be None, which would break
setup_and_run. So lets check if there is a function to be executed
first. If there isn't, this is just a noop on the hands of the debugger.