Changing memescript source file language structure.
I was looking for having all functions (and methods) starting at column
0 so the identation is not screwed in the idez or file. I then took a
leap on structuring the module source files as section.

The top level sections are: .license, .meta, .preamble and .code

* .license is free text for the licensing text.

* .meta will be used in the future for the module's meta information.

* .preamble informs the module parameters, default values for them and
  "aliases".

* .code is where module functions and classes are defined.

Finally, a smaller change was in the syntax for opening and closing
class declarations: there are no brackets now, we just end the
declaration with "end", ala ruby. This was chosen because using brackets
would make the class declaration end with two lines each with "}": one
for closing the last method and the other for closing the class
declaration -- which is much uglier.