Controlling debugger execution on exception.
A stack of flags keep track of "catches" associated with the current
running code. Those flags indicate if the debugger should be fired or
not (ie, when there is no protection for the running
code). exception_unprotected() is used to "unflag" a certain function,
so that any code unprotected in it that raises will fire the debugger.