Compiling structures with dicts key'd by symbols
Also, added a hacky mm_list_index_of to handle string comparision.

The reason is that d["foo"] would fail for a dictionary with key "foo"
but whose object string was created elsewhere (say, another module, or
computed dynamically). Since we will need to access these dicts when
developing in the near future, this provisional impl. should suffice for
now.