Context creation methods as primitives in core.
When stepping into closure declarations, we ended up stepping into the
methods that create closure.

Also, debugger is crashing when accessing `frame.fp` or `frame.ip`,
which are pointers but not oop, thus the crashes. I made provisional
changes so the code (both in core.mm and debugger) do not access these
pointers. I think an improved solution would be to return frame.fp and
frame.ip as Number instances.