Being more careful about exceptions in mmobj.
Because mmobj is used a lot by prims, we would like to check for type
errors and raise(). However, mmobj is also used by internal routines in
the vm/process and raise()ing during these routines might put the system
in an unstable state. Rather, we might be better off assert()ing.

To accomplish that, callers of mmobj that should assert() if something
goes wrong pass an extra true argument