Fix closure parsing of last expr/stmt.
Previously, the parser was making a distinction of return value
depending on the last expr being terminated with ';' (returning null) or
not (returning the value of the expr).

Now, closures return the value of the last expr always, or null if
there's no body or last command is a statement, regardless of ';'
presence.