bindgen: Add parser & generator for C function signatures
This is the first step towards a system that automatically generates
bindings of C functions to MemeTalk. This prototype will read the
definitions of a separate file and generate the C++ primitives for
them.

For the input `//sys int close(int fd)`, bindgen will output the
following code:

  static int prim_close (Process* proc) {
    oop oop_fd = proc->get_arg(0);
    if (!(proc->mmobj()->mm_object_vt(oop_fd) == proc->vm()->get_prime("Integer"))) {
      proc->raise("TypeError", e);
    }
    int fd = untag_small_int(oop_fd);

    /* Call to the underlying function */
    int output = close (fd);
    oop oop_output = tag_small_int(oop_output);
    proc->stack_push(oop_output);
    return 0;
  }

The only currently supported types are `int` & `char` it will be
extended to more of the built-in types and to custom structs. Both
`const` and `unsigned` are also fully supported in parameters and in
the return type.

Pointers are parsed and sent over to the translator that doesn't know
too much about them yet. It currently doesn't support pointers of
pointers. It wasn't implemented yet because it might touch a little
bit the initialization of structs & output parameters.

There are three mappings that must be updated for extending the
translator (besides the grammar):

 * Fun.genChkArg()
 * Fun.genUnwrapArg()
 * Fun.genFunCall()