bindgen: Generate register() function & multiple small fixes
The `_prims.c` file now contains the function `init_primitives` that
takes a pointer to a `VM` object and registers each primitive. Nothing
is really going to execute this function until the primitive system is
changed to accept dynamically loaded extensions (which might happen
soon).

The other minor fixes were:

 * Print out as many pointers as defined
 * Fix syntax error caused by extra parameter after unwrapping strings
 * Normalize names of parameters passed to the function being called