bindgen: Rewrite how code C is generated
The C generator needs to concatenate lots of strings. This patch
introduces lots of very small functions prefixed with `c_` that can be
combined and represent the code that needs to be generated with code
that looks like the C output we want.

It's also way more productive to write tests with these `c_` prefixed
functions than manually.