Trying to clean up a bit error api.
All .net functions returning -1 in the unerlying API now raise
NetworkError. Because often sockets are a result of a chain of calls to
getaddrinfo(), socket(), bind() and connect() over the addrinfo linked list, I
also added List.first_success() so that we can iterate over addrinfos and call
each of these functions; upon NetworkError, the list iterate over the next
addrs. So the usage seems cleaner.